requirements.txt
asgiref~=3.7.2
certifi~=2023.7.22
charset-normalizer~=3.2.0
colorclass~=2.2.2
Django~=4.2.4
django-extensions~=3.2.3
django-rest-framework-nested-resource~=1.3.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
docopt~=0.6.2
drf-nested-routers~=0.93.4
drf-yasg~=1.21.7
idna~=3.4
inflection~=0.5.1
packaging~=23.1
Pillow~=10.0.0
pip-upgrader~=1.4.15
PyJWT~=2.8.0
pytz~=2022.6
PyYAML~=6.0.1
requests~=2.31.0
routers~=0.10.1
sqlparse~=0.4.4
terminaltables~=3.1.10
tzdata~=2023.3
uritemplate~=4.1.1
urllib3~=2.0.4
